\name{MaxSkewThree}
\alias{MaxSkewThree}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{MaxSkew: skewness-based projection pursuit

}
\description{Finds Orthogonal Data Projections with Maximal Skewness for Trivariate Random Vectors
}
\usage{
MaxSkewThree(data,iterations)
}
\arguments{
  \item{data}{Data matrix where rows and columns represent units and variables. 
}
  \item{iterations}{Number of required iterations.

}
}
\details{It is an internal function called by MaxSkew

}







\author{Cinzia Franceschini and Nicola Loperfido
}

\references{de Lathauwer L., de Moor B.and Vandewalle J. (2000). Onthebestrank-1andrank-(R_{1},R_{2},...R_{N}) approximation of high-order tensors. SIAM Jour. Matrix Ana. Appl. 21, 1324-1342.

Loperfido, N. (2010). Canonical Transformations of Skew-Normal Variates. Test 19, 146-165.

Loperfido, N. (2013). Skewness and the Linear Discriminant Function. Statistics and Probability Letters 83, 93-99.

Malkovich, J.F. and Afifi, A.A. (1973). On Tests for Multivariate Normality. J. Amer. Statist. Ass. 68, 176-179.
%% ~put references to the literature/web site here ~
}

